Separate engine error related definitions from other definitions.
For clarity.
