+++ src/engine/errors.rs
@@ -1,0 +1,52 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+use std::io;
+
+use nix;
+
+#[derive(Debug)]
+pub enum ErrorEnum {
+    Ok,
+    Error(String),
+
+    AlreadyExists(String),
+    Busy(String),
+    Invalid(String),
+    NotFound(String),
+}
+
+impl ErrorEnum {
+    pub fn get_error_string(&self) -> String {
+        match *self {
+            ErrorEnum::Ok => "Ok".into(),
+            ErrorEnum::Error(ref x) => format!("{}", x),
+            ErrorEnum::AlreadyExists(ref x) => format!("{} already exists", x),
+            ErrorEnum::Busy(ref x) => format!("{} is busy", x),
+            ErrorEnum::Invalid(ref x) => format!("{}", x),
+            ErrorEnum::NotFound(ref x) => format!("{} is not found", x),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum EngineError {
+    Stratis(ErrorEnum),
+    Io(io::Error),
+    Nix(nix::Error),
+}
+
+pub type EngineResult<T> = Result<T, EngineError>;
+
+impl From<io::Error> for EngineError {
+    fn from(err: io::Error) -> EngineError {
+        EngineError::Io(err)
+    }
+}
+
+impl From<nix::Error> for EngineError {
+    fn from(err: nix::Error) -> EngineError {
+        EngineError::Nix(err)
+    }
+}
