+++ src/engine/mod.rs
@@ -4,12 +4,13 @@
 
 pub use self::engine::Dev;
 pub use self::engine::Engine;
-pub use self::engine::EngineError;
-pub use self::engine::EngineResult;
-pub use self::engine::ErrorEnum;
 pub use self::engine::Filesystem;
 pub use self::engine::Pool;
 pub use self::engine::RenameAction;
+
+pub use self::errors::EngineError;
+pub use self::errors::EngineResult;
+pub use self::errors::ErrorEnum;
 
 #[macro_use]
 mod macros;
@@ -18,3 +19,4 @@ pub mod sim_engine;
 pub mod strat_engine;
 
 mod engine;
+mod errors;
